requirements.txt
discord.py
pytz
pandas
matplotlib
sqlalchemy
